requirements.txt
# MuseTAG Dependencies

# Web Framework
streamlit>=1.28.0

# HTTP Requests
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Audio Processing (optional, for duration detection)
mutagen>=1.47.0

# Audio Truncation (optional, for large files)
pydub>=0.25.1

# Note: pydub requires ffmpeg to be installed for audio processing
# On macOS: brew install ffmpeg
# On Ubuntu: sudo apt-get install ffmpeg
